chore: move utils out of apps
