+++ test/integration/02-user-wallet/02-receive-onchain.spec.ts
@@ -13,8 +13,7 @@ import { baseLogger } from "@services/logger"
 import { User } from "@services/mongoose/schema"
 import { getTitle } from "@services/notifications/payment"
 
-import { sleep } from "@app/utils"
-
+import { sleep } from "src/utils"
 import {
   amountAfterFeeDeduction,
   bitcoindClient,
