+++ test/integration/02-user-wallet/02-send-onchain.spec.ts
@@ -18,7 +18,7 @@ import { Transaction } from "@services/mongoose/schema"
 import { getTitle } from "@services/notifications/payment"
 import last from "lodash.last"
 
-import { sleep } from "@app/utils"
+import { sleep } from "src/utils"
 
 import {
   bitcoindClient,
