+++ test/integration/servers/graphql-main-server/no-auth-requests.spec.ts
@@ -1,5 +1,5 @@
 import { createHttpTerminator } from "http-terminator"
-import { sleep } from "@app/utils"
+
 import { yamlConfig, JWT_SECRET } from "@config/app"
 import { createTestClient } from "apollo-server-integration-testing"
 import { startApolloServerForCoreSchema } from "@servers/graphql-main-server"
@@ -7,6 +7,8 @@ import * as jwt from "jsonwebtoken"
 
 import USER_REQUEST_AUTH_CODE from "./mutations/user-request-auth-code.gql"
 import USER_LOGIN from "./mutations/user-login.gql"
+
+import { sleep } from "src/utils"
 
 import { clearAccountLocks, clearLimiters } from "test/helpers"
